Set default client id for new response, event out, and request nodes
If only one client config node has been defined for a flow, that config
node's id is now set as the client id when a new response, event out,
or request node is added to the flow. Previously, a user had to at least
edit the new node description in order for the client id to be set for
the node - even though no other fields were required in order for the
node to be valid.