Initial examples and auto-select first client on node import
This commit adds a couple of initial example json flows files - basic
and event. To enable the flows to be imported without requiring the user
to open each node in the UI and select an associated client config node,
this commit also adds some logic which auto associates the first available
client node if the node is not already associated with a valid client
node. The association only occurs if a single client node is available.
If more than one client node is available, a user presumably would want
to specifically select the client which should be used for each node.

This commit also includes a new script, remove-config-node, which can
be run to remove all client nodes from a flows file. This script would
presumably be run before distributing an examples file in order to
ensure that importing the examples file into a flow does not
unintentionally drag in extra client configuration nodes that a user
would not want.