Store converted value for payload under dxlMessage/Request/Response keys
Previously, the value converted from the raw payload in a message (an
object or string) was only stored for the msg.payload key. The value
stored for the payload key under the dxlMessage, dxlRequest, and
dxlResponse keys was the original binary Buffer received for the
message. With the changes in this commit, the converted value is stored
in each location where the payload is stored into the flow msg.