Use new dxlError msg property to populate errorCode in error response
Previously, the errorCode set in an error response was parsed out of the
contents of the error message string, with the expectation that it would
appear inside of parentheses at the end of the string. This commit
reserves a separate property in the flow msg, dxlError.code, for setting
the errorCode, eliminating the need to concatenate the error code into
the message.