Update dxl client ids for existing nodes when config saved/deleted
Previously, when a DXL client configuration was created in the flow
editor, any existing nodes which have a DXL client field were not
automatically associated with the new configuration node. With the
changes in this commit, the configuration node would be associated with
the pre-existing nodes.

Note that the association only occurs when adding a client node via the
edit box on a consuming node. If the configuration node is imported
into the flow, the association does not automatically occur. In this
case, the user would need to at least open and save/update the
configuration node in the UI in order for the association to be
automatically applied to consuming nodes.

If two configuration nodes are present and one of the nodes is removed,
the remaining node is automatically associated to all consuming nodes.