Add extractProperty/ies NodeUtils functions
This commit adds a couple of utility functions - extractProperty and
extractProperties - which can be used for removing one or more
properties from an object and obtaining the values for those properties.
These are generally expected to be used by nodes for reading properties
from an incoming msg and removing the properties from the outgoing msg
to avoid polluting the msg that might be received by downstream nodes of
the same type.