Populate dxlError.code in response node from errorResponse.errorCode
Previously, the `dxlError.code` property was always populated onto the
flow message from the corresponding `code` property on the
`MessageError` received in the response for a failed request. In later
OpenDXL JavaScript client releases, the `MessageError` class (changed
to `RequestError`) always populates its `code` property as a string
value and is empty for custom error codes.

This commit changes the logic for populating the `dxlError.code`
property to use the string-based `code` if non-empty but otherwise
falls back to the numeric `errorCode` in the `dxlErrorResponse` included
in the `RequestError` instance.