Add confirmation dialog when provisioned files already exist
Previously, if a user were to attempt a provision operation for a config
directory in which the provision files already exist, the operation
would proceed without warning the user that existing files on the server
would be replaced. With the changes in this commit, the user would see a
confirmation dialog and have to press a 'Confirm provision' button in
order to proceed with replacing the files on the server. The
confirmation dialog is not displayed if the target config directory does
not already have one or more of the files that would be stored by the
provisioning operation.