Use custom port when new provisioning operation performed
Previously, if a user specified a custom value for the 'port' field in a
new provisioning operation, the custom value was ignored and the default
port, 8883, was always used instead. With the changes in this commit,
the port specified by the user is now passed along properly for the
provisioning operation.